feat(exams): Add functionality to send schedule
This commit adds the functionality to send a schedule for exams. The code now includes a new state variable `url` and `data` in the Exams component. The `sendSchedule` function has been updated to set the `data` state with the necessary information for scheduling exams. Additionally, the `setUrl` function is called to update the URL for sending the schedule.

The commit also includes changes to handle response and error messages when sending the schedule. If successful, a confirmation modal is displayed with instructions on how to access the app for consultation. If there is an error, an error modal is displayed with a message asking users to try again.

Lastly, the "Finalizar" button is now disabled if no exams are selected.

Closes #123