feat(radio): Add support for displaying titles
This commit adds support for displaying titles in the Radio component. The `Radio` component now accepts a new prop `titles`, which is an array of strings representing the titles for each radio option. If `titles` is provided, the radio options will be displayed in a column layout with the title on top and the text below. Additionally, some CSS classes have been added to style the new layout.

The changes also include updates to the usage of the `Radio` component in the `Unit` page, where it now uses the `titles` prop to display unit numbers as titles and addresses as text.

Closes #123